refactor(app): extract mouse-driven scrollbar input (phase 4b)
Move handle_mouse_scroll, handle_horizontal_scroll, handle_scrollbar_jump,
handle_scrollbar_drag_relative, handle_composite_scrollbar_jump,
handle_composite_scrollbar_drag_relative out of input.rs into a new
app/scrollbar_input.rs sibling. ~636 lines.

Mouse-wheel scrolling, click-to-jump, and click-and-drag scrollbar
behaviour for both regular and composite (diff-view) scrollbars.
Pure scrollbar math already lives in app/scrollbar_math.rs (phase 1i);
these methods are the side-effecting drivers.

input.rs goes from 2941 -> 2304 lines (-637). 2026 tests still pass.
