+++ crates/fresh-editor/src/app/mod.rs
@@ -44,6 +44,7 @@ mod recovery_actions;
 mod regex_replace;
 mod render;
 mod scroll_sync;
+mod scrollbar_input;
 mod scrollbar_math;
 mod search_ops;
 mod search_scan;
